Added alert

+++ src/components/Dashboard/Card.jsx
@@ -1,20 +1,49 @@
 /* eslint-disable react/prop-types */
 
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const Card = ({ product, onDelete }) => {
     const { id, title, price, image, category } = product;
 
     const handleDelete = async () => {
-        await fetch(`http://localhost:3000/products/${id}`, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(() => {
-                onDelete(id);
-        })
+
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                try {
+                    await fetch(`http://localhost:3000/products/${id}`, {
+                        method: 'DELETE',
+                    })
+                        .then(res => res.json())
+                        .then(() => {
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+
+                            onDelete(id);
+                        })
+
+                } catch (error) {
+                    Swal.fire({
+                        title: "Error!",
+                        text: "There was a problem deleting your file.",
+                        icon: "error"
+                    });
+                }
+            }
+        });
     }
-    
+
     return (
         <>
             <div className="card card-compact bg-base-100 shadow-xl">
